pyproject.toml
[project]
name = "dafuweng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.6",
    "openai>=2.15.0",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
    "rapidocr-onnxruntime>=1.4.4",
]
